# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OenoFile Express API",
    "description": "API service for the OenoFile web and mobile app",
    "contact": {
      "email": "ccarson17@gmail.com"
    },
    "version": "1.0"
  },
...
  "paths": {
...
    "/api/v1/user/{id}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Finds a user by its ID",
        "description": "Will return users matching your ownerId",
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID to update",
            "required": true,
            "explode": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          }
        }
      },
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Update an existing user",
        "description": "Update an existing user by Id",
        "operationId": "putUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID to update",
            "required": true,
            "explode": true
          }
        ],
        "requestBody": {
          "description": "Update an existing user in the DB (full update)",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad request data"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "patch": {
        "tags": [
          "user"
        ],
        "summary": "Update an existing user",
        "description": "Update an existing user by Id",
        "operationId": "patchUser",
        "requestBody": {
          "description": "Update an existing user in the DB (partial update)",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID to update",
            "required": true,
            "explode": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad request data"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "Delete an existing user",
        "description": "Delete an existing user by Id",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID to delete",
            "required": true,
            "explode": true
          }
        ],
        "responses": {
          "204": {
            "description": "Successful operation (no content)"
          },
          "400": {
            "description": "Bad request data"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "User": {
        "required": [
          "userName"
        ],
        "type": "object",
        "properties": {
          "userName": {
            "type": "string",
            "example": "JoeSmith"
          }
        },
        "xml": {
          "name": "user"
        }
      }
    },
...
    "securitySchemes": { }
  }
}